feat: consolidate navigation into global NavBar and allow decimal expected games
Move ThemeToggle and AuthButton into the global NavBar so all pages
(home, league, draft, rankings) share consistent navigation. Remove
duplicate page-level headers. Also change expected games inputs from
integer dropdowns to decimal number inputs.
