+++ app/components/nav-bar.tsx
@@ -1,36 +1,17 @@
-'use client';
-
-import Link from 'next/link';
-import { usePathname } from 'next/navigation';
-
-const navLinks = [
-  { href: '/', label: 'Home' },
-  { href: '/rankings/ncaa', label: 'Rankings' },
-];
+import AuthButtonServer from '../auth-button-server';
+import ThemeToggle from '../theme-toggle';
+import NavBarClient from './nav-bar-client';
 
 export default function NavBar() {
-  const pathname = usePathname();
-
   return (
     <nav className="border-b border-border bg-card">
       <div className="max-w-6xl mx-auto px-4">
-        <div className="flex items-center h-12 gap-6">
-          <Link href="/" className="font-bold text-primary text-lg mr-4">
-            Fanteasy
-          </Link>
-          {navLinks.map((link) => (
-            <Link
-              key={link.href}
-              href={link.href}
-              className={`text-sm transition-colors ${
-                pathname === link.href || pathname?.startsWith(link.href + '/')
-                  ? 'text-foreground font-medium'
-                  : 'text-muted-foreground hover:text-foreground'
-              }`}
-            >
-              {link.label}
-            </Link>
-          ))}
+        <div className="flex items-center justify-between h-12">
+          <NavBarClient />
+          <div className="flex items-center gap-3">
+            <ThemeToggle />
+            <AuthButtonServer />
+          </div>
         </div>
       </div>
     </nav>
